feat: full-width order detail; unify payment, attachments, courier
Restructure the order detail page from a 2/3 + 1/3 split into a single
full-width vertical stack (right column removed):

- Payment: pull the money summary out of line items and merge it with
  payment status/method/paid-at and the mark-as-paid form into one
  section — surfaces method + paid-at, which the page never showed.
- Attachments: consolidate drop-off + pickup photos into one card,
  2-col on desktop; reuses PhotoUploadDialog/PhotoLightbox.
- Courier: move into the identity-strip overflow menu as a dialog;
  show the assigned courier in the meta line.
- Add shared OrderSectionHeader; retire order-courier-card,
  order-payment-card, order-dropoff-photo-card, order-pickup-history-card.

Document the drop-off photo rule (grill-with-docs): required for every
order incl. product-only, attached best-effort after commit, detail
page is the recovery path.

- CONTEXT.md: sharpen the drop-off photo glossary entry.
- ADR-0014: record the required/best-effort policy + rejected options.
- seed: photo for all order types (~10% miss to exercise recovery),
  was wrongly service-gated.
