+++ apps/web/src/features/orders/components/order-identity-strip.tsx
@@ -1,6 +1,7 @@
 import {
 	DotsThreeVerticalIcon,
 	LinkSimpleIcon,
+	TruckIcon,
 	WarningCircleIcon,
 } from "@phosphor-icons/react";
 import { toast } from "sonner";
@@ -16,6 +17,7 @@ import {
 import { OpenComplaintForm } from "@/features/complaints/components/open-complaint-form";
 import { useOpenComplaintMutation } from "@/features/complaints/hooks/useComplaintMutations";
 import { CancelOrderForm } from "@/features/orders/components/cancel-order-form";
+import { OrderCourierForm } from "@/features/orders/components/order-courier-form";
 import { OrderPickupEventDialog } from "@/features/orders/components/order-pickup-event-dialog";
 import { RefundOrderForm } from "@/features/orders/components/refund-order-form";
 import {
@@ -98,6 +100,24 @@ export const OrderIdentityStrip = ({
 		});
 	};
 
+	const openCourierDialog = () => {
+		openDialog({
+			title: "Set courier",
+			description:
+				"Assign a courier to collect this order, or leave as walk-in.",
+			contentClassName: "sm:max-w-md",
+			content: () => (
+				<OrderCourierForm
+					closeDialog={closeDialog}
+					currentCourierId={
+						detail.collected_by ? String(detail.collected_by) : ""
+					}
+					orderId={orderId}
+				/>
+			),
+		});
+	};
+
 	const openCancelOrderDialog = () => {
 		openDialog({
 			title: "Cancel order",
@@ -167,6 +187,7 @@ export const OrderIdentityStrip = ({
 
 	const hasMenu =
 		Boolean(trackingUrl) ||
+		gates.canManageCourier ||
 		gates.canCancelOrder ||
 		gates.canRefundWholeOrder ||
 		gates.canOpenComplaint;
@@ -176,6 +197,7 @@ export const OrderIdentityStrip = ({
 		detail.customer?.phone_number,
 		detail.store?.name,
 		formatOrderDateTime(detail.created_at),
+		detail.collectedBy ? `Courier: ${detail.collectedBy.name}` : null,
 	]
 		.filter(Boolean)
 		.join(" · ");
@@ -238,11 +260,17 @@ export const OrderIdentityStrip = ({
 											/>
 										}
 									/>
-									<DropdownMenuContent align="end">
+									<DropdownMenuContent align="end" className="w-40">
 										{trackingUrl ? (
 											<DropdownMenuItem onClick={handleCopyTrackingLink}>
 												<LinkSimpleIcon className="size-4" />
 												Copy tracking link
+											</DropdownMenuItem>
+										) : null}
+										{gates.canManageCourier ? (
+											<DropdownMenuItem onClick={openCourierDialog}>
+												<TruckIcon className="size-4" />
+												Set courier
 											</DropdownMenuItem>
 										) : null}
 										{gates.canOpenComplaint ? (
