+++ apps/web/src/features/orders/components/order-section-header.tsx
@@ -1,0 +1,7 @@
+import type { ReactNode } from "react";
+
+export const OrderSectionHeader = ({ children }: { children: ReactNode }) => (
+	<div className="flex items-center justify-between px-4 py-2.5">
+		<p className="text-foreground text-sm font-semibold">{children}</p>
+	</div>
+);
